feat(components-pipeline-1): move header to molecules
