+++ design-system/react/src/molecules/header/Header.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { Button } from '../src/atoms/button/Button';
+import { Button } from '../../atoms/button/Button';
 import './header.css';
 
 export const Header = ({ user, onLogin, onLogout, onCreateAccount }) => (
