+++ design-system/react/stories/Header.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Header } from './Header';
+import { Header } from '../src/molecules/header/Header';
 
 export default {
   title: 'Example/Header',
